# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexPill API
  version: 1.0.0
  description: Versioned mobile API. All protected routes require a Laravel Sanctum bearer token.
servers:
  - url: https://api.example.com/api/v1
  - url: http://127.0.0.1:8000/api/v1
...
paths:
...
  /auth/otp/request:
    post:
      tags: [Auth]
      operationId: requestOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, purpose]
              properties:
                email: { type: string, format: email }
                purpose: { type: string, enum: [login, register] }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /auth/otp/verify:
    post:
      tags: [Auth]
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, purpose, device_name]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: '^[0-9]{6}$' }
                purpose: { type: string, enum: [login, register] }
                device_name: { type: string }
                name: { type: string }
                timezone: { type: string, example: Asia/Kolkata }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...
  responses:
    Success:
      description: Successful NexPill response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
  schemas:
    Envelope:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        data: {}
        errors: {}
...